requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.6
django-environ==0.4.5
django-taggit==1.5.1
pytz==2021.1
sqlparse==0.4.1
